loader: fix mismatched type for inline assembly

+++ loader/src/aarch64/init.c
@@ -137,7 +137,7 @@ typedef void (*sel4_entry)(
 void arch_jump_to_kernel(int logical_cpu)
 {
     /* seL4 always expects the current logical CPU number in TPIDR_EL1 */
-    asm volatile("msr TPIDR_EL1, %0" :: "r"(logical_cpu));
+    asm volatile("msr TPIDR_EL1, %0" :: "r"((uint64_t) logical_cpu));
     asm volatile("isb" ::: "memory");
 
     ((sel4_entry)(loader_data->kernel_entry))(
